+++ packages/uniforms-material/src/ListAddField.js
@@ -12,6 +12,7 @@ const ListAdd = ({
     value,
     ...props
 }) => {
+    const limitNotReached = !disabled && !(parent.maxCount <= parent.value.length);
 
     return (
         <RaisedButton